Remove page id from extension script injection permission.
It already throws out pending requests when a new page loads, so the page id checks were redundant. Also, request ids are 64 bits, so fix them everywhere.

BUG=371151
TEST=everything still works

Review URL: https://codereview.chromium.org/362413003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281183 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d2056001032e3c896409d4cad2f02a9627d38920

+++ extensions/renderer/script_injection_manager.cc
@@ -286,7 +286,7 @@ void ScriptInjectionManager::HandleExecuteCode(
   }
 }
 
-void ScriptInjectionManager::HandlePermitScriptInjection(int request_id) {
+void ScriptInjectionManager::HandlePermitScriptInjection(int64 request_id) {
   ScopedVector<ScriptInjection>::iterator iter =
       pending_injections_.begin();
   for (; iter != pending_injections_.end(); ++iter) {
@@ -295,6 +295,11 @@ void ScriptInjectionManager::HandlePermitScriptInjection(int request_id) {
   }
   if (iter == pending_injections_.end())
     return;
+
+  // At this point, because the request is present in pending_injections_, we
+  // know that this is the same page that issued the request (otherwise,
+  // RVOHelper's DidStartProvisionalLoad callback would have caused it to be
+  // cleared out).
 
   scoped_ptr<ScriptInjection> injection(*iter);
   pending_injections_.weak_erase(iter);
